# serving/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Commerce Serving API",
    "description": "Read-only BI/API serving surface over Trino + Iceberg.",
    "version": "1.0.0"
  },
  "paths": {
    "/health/live": {
      "get": {
        "summary": "Live",
        "description": "Liveness：只证明 FastAPI 进程可响应，不证明 Trino / Iceberg 已可查询。",
        "operationId": "live_health_live_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "summary": "Ready",
        "description": "Readiness：通过 Repository 执行最小 Trino 查询；失败统一返回 HTTP 503。\n\n工程边界：Ready 只证明查询入口可用，不证明某个业务日 Serving 分区已经 Fresh。",
        "operationId": "ready_health_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/executive/daily": {
      "get": {
        "summary": "Executive Daily",
        "description": "返回一个业务日按 Region 预计算的 Executive Metrics。",
        "operationId": "executive_daily_api_v1_executive_daily_get",
        "parameters": [
          {
            "name": "business_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Business Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyExecutiveRow"
                  },
                  "title": "Response Executive Daily Api V1 Executive Daily Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/regions/{region}/daily": {
      "get": {
        "summary": "Region Daily",
        "description": "返回一个 Region 的日粒度 Serving Rows；不在 API 层重新聚合 AOV / Net Sales 公式。",
        "operationId": "region_daily_api_v1_regions__region__daily_get",
        "parameters": [
          {
            "name": "region",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Region"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyExecutiveRow"
                  },
                  "title": "Response Region Daily Api V1 Regions  Region  Daily Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DailyExecutiveRow": {
        "properties": {
          "business_date": {
            "type": "string",
            "format": "date",
            "title": "Business Date"
          },
          "region": {
            "type": "string",
            "title": "Region"
          },
          "gross_sales": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gross Sales"
          },
          "sales_before_reversal": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sales Before Reversal"
          },
          "net_sales": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Sales"
          },
          "order_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Count"
          },
          "average_order_value": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Order Value"
          }
        },
        "type": "object",
        "required": [
          "business_date",
          "region"
        ],
        "title": "DailyExecutiveRow",
        "description": "Executive Serving 的稳定 JSON Row；字段与 ``bi_daily_executive`` Contract 一一对应。\n\nDecimal 保留指标金额精度；模型只做响应校验，不承担二次计算。"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "HealthResponse",
        "description": "健康检查最小响应合同，供容器/调用方区分 live 与 ready 结果。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}